+++ jazzy_ws/src/isaacsim/launch/run_isaacsim.launch.py
@@ -1,9 +1,17 @@
-## Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
-## NVIDIA CORPORATION and its licensors retain all intellectual property
-## and proprietary rights in and to this software, related documentation
-## and any modifications thereto.  Any use, reproduction, disclosure or
-## distribution of this software and related documentation without an express
-## license agreement from NVIDIA CORPORATION is strictly prohibited.
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from launch import LaunchDescription
 from ament_index_python.packages import get_package_share_directory
@@ -20,7 +28,7 @@ launch_args = [
     
     DeclareLaunchArgument('install_path', default_value='', description='If Isaac Sim is insalled in a non-default location, provide a specific path to Isaac Sim installation root folder. (If defined, "version" parameter will be ignored)'),
     
-    DeclareLaunchArgument('use_internal_libs', default_value='false', description='Set to true if you wish to use internal ROS libraries shipped with Isaac Sim.'),
+    DeclareLaunchArgument('use_internal_libs', default_value='true', description='Set to true if you wish to use internal ROS libraries shipped with Isaac Sim.'),
     
     DeclareLaunchArgument('dds_type', default_value='fastdds', description='Set to "fastdds" or "cyclonedds" (Cyclone only supported for ROS 2 Humble and ROS 2 Jazzy) to run Isaac Sim with a specific dds type.'),
     
@@ -32,40 +40,18 @@ launch_args = [
     
     DeclareLaunchArgument('ros_distro', default_value='jazzy', description='Provide ROS version to use. Only Humble and Jazzy is supported.'),
     
-    DeclareLaunchArgument('ros_installation_path', default_value='', description='If ROS is installed in a non-default location (as in not under /opt/ros/), provide the path to your main setup.bash file for your ROS install. (/path/to/custom/ros/install/setup.bash)'),
+    DeclareLaunchArgument('ros_installation_path', default_value='', description='Comma-separated list of ROS installation paths. If ROS is installed in a non-default location (as in not under /opt/ros/), provide the path to your main setup.bash file for your ROS install. (/path/to/custom/ros/install/setup.bash). Similarly add the path to your local_setup.bash file for your workspace installation. (/path/to/custom_ros_workspace/install/local_setup.bash)'),
 
     DeclareLaunchArgument('headless', default_value='', description='Set to "native" or "webrtc" to run Isaac Sim with different headless modes, if left blank, Isaac Sim will run in the regular GUI workflow. This parameter can be overridden by "standalone" parameter.'),
 
     DeclareLaunchArgument('custom_args', default_value='', description='Add any custom Isaac Sim args that you want to forward to isaac-sim.sh during run time.'),
 
+    DeclareLaunchArgument('exclude_install_path', default_value='', description='Comma-separated list of installation paths to exclude from LD_LIBRARY_PATH, PYTHONPATH, and PATH environment variables.'),
+
 ]
 
-# List of parameters to check
-parameters_to_check = [
-    'version', 'install_path', 'use_internal_libs', 'dds_type',
-    'standalone', 'play_sim_on_start', 'ros_distro', 'ros_installation_path', 'gui'
-]
 
 def launch_setup(context):
-
-    def add_parameter_argument(param_name, command_list):
-        param_value = LaunchConfiguration(param_name).perform(context) # Here you'll get the runtime config value
-
-        if param_value != '':
-            if param_value == 'true':
-                return command_list + ['--' + param_name.replace('_', '-')]
-            elif param_value == 'false':
-                return command_list
-            else:
-                return command_list + ['--' + param_name.replace('_', '-'), param_value,]
-        else:
-            return command_list
-
-    bash_command = []
-
-    # Add parameters to the command list if they are set
-    for param in parameters_to_check:
-        bash_command = add_parameter_argument(param, bash_command)
         
     # Run isaac sim as a ROS2 node with default parameters. Parameters can be overridden here or via launch arguments from other launch files. 
     isaacsim_node = Node(
@@ -82,7 +68,8 @@ def launch_setup(context):
             'ros_distro': LaunchConfiguration('ros_distro'),
             'ros_installation_path': LaunchConfiguration('ros_installation_path'),
             'headless': LaunchConfiguration('headless'),
-            'custom_args': LaunchConfiguration('custom_args')
+            'custom_args': LaunchConfiguration('custom_args'),
+            'exclude_install_path': LaunchConfiguration('exclude_install_path')
         }]
     )
     return [isaacsim_node]